requirements.txt
asgiref==3.5.2
Django==4.1.1
django-db==0.0.7
django-filter==22.1
djangorestframework==3.13.1
joblib==1.2.0
Markdown==3.4.1
numpy==1.23.3
pandas==1.5.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.2
xgboost==1.5.0
gunicorn==19.9.0